# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/v2/albums": {
            "get": {
                "description": "Retrieves a list of albums with optional sorting, pagination, and search",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "albums"
                ],
                "summary": "Get a list of albums",
                "parameters": [
                    {
                        "type": "string",
                        "default": "id",
                        "description": "Column to order by",
                        "name": "order_by",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "asc",
                        "description": "Sort direction (asc or desc)",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "20",
                        "description": "Number of records to return",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "0",
                        "description": "Offset for pagination",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search term",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.AlbumsWithTotalCount"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Creates a new album with the provided details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "albums"
                ],
                "summary": "Create a new album",
                "parameters": [
                    {
                        "description": "Album details",
                        "name": "album",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateAlbum"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Album"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    }
                }
            }
        },
        "/api/v2/albums/{id}": {
            "get": {
                "description": "Retrieves details of a specific album by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "albums"
                ],
                "summary": "Get album details",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Album ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.AlbumWithImages"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Updates the details of a specific album by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "albums"
                ],
                "summary": "Update album details",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Album ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated album details",
                        "name": "album",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateAlbum"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.CreateAlbum"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Deletes a specific album by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "albums"
                ],
                "summary": "Delete an album",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Album ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "integer"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    }
                }
            }
        },
        "/api/v2/albums/{id}/cover/{imageName}": {
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Sets a specific image as the cover for an album by album ID and image name",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "albums"
                ],
                "summary": "Set album cover image",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Album ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Image Name",
                        "name": "imageName",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    }
                }
            }
        },
        "/api/v2/albums/{id}/images": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Uploads one or more images to the specified album",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "albums"
                ],
                "summary": "Upload images to an album",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Album ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "file"
                        },
                        "collectionFormat": "multi",
                        "description": "Images to upload",
                        "name": "images",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    }
                }
            }
        },
        "/api/v2/albums/{id}/images/{imageName}": {
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Deletes a specific image from an album by album ID and image name",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "albums"
                ],
                "summary": "Delete an image from an album",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Album ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Image Name",
                        "name": "imageName",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "internal.LocalTime": {
            "type": "object",
            "properties": {
                "time.Time": {
                    "type": "string"
                }
            }
        },
        "models.Album": {
            "type": "object",
            "properties": {
                "created_at": {
                    "$ref": "#/definitions/internal.LocalTime"
                },
                "description_en": {
                    "type": "string"
                },
                "description_no": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name_en": {
                    "type": "string"
                },
                "name_no": {
                    "type": "string"
                },
                "updated_at": {
                    "$ref": "#/definitions/internal.LocalTime"
                },
                "year": {
                    "type": "integer"
                }
            }
        },
        "models.AlbumWithImages": {
            "type": "object",
            "properties": {
                "created_at": {
                    "$ref": "#/definitions/internal.LocalTime"
                },
                "description_en": {
                    "type": "string"
                },
                "description_no": {
                    "type": "string"
                },
                "event": {
                    "$ref": "#/definitions/models.EventAlbum"
                },
                "id": {
                    "type": "integer"
                },
                "images": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "name_en": {
                    "type": "string"
                },
                "name_no": {
                    "type": "string"
                },
                "updated_at": {
                    "$ref": "#/definitions/internal.LocalTime"
                },
                "year": {
                    "type": "integer"
                }
            }
        },
        "models.AlbumsWithTotalCount": {
            "type": "object",
            "properties": {
                "created_at": {
                    "$ref": "#/definitions/internal.LocalTime"
                },
                "description_en": {
                    "type": "string"
                },
                "description_no": {
                    "type": "string"
                },
                "event": {
                    "$ref": "#/definitions/models.EventAlbum"
                },
                "id": {
                    "type": "integer"
                },
                "images": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "name_en": {
                    "type": "string"
                },
                "name_no": {
                    "type": "string"
                },
                "total_count": {
                    "type": "integer"
                },
                "updated_at": {
                    "$ref": "#/definitions/internal.LocalTime"
                },
                "year": {
                    "type": "integer"
                }
            }
        },
...
        "models.CreateAlbum": {
            "type": "object",
            "properties": {
                "created_at": {
                    "$ref": "#/definitions/internal.LocalTime"
                },
                "description_en": {
                    "type": "string"
                },
                "description_no": {
                    "type": "string"
                },
                "event_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "name_en": {
                    "type": "string"
                },
                "name_no": {
                    "type": "string"
                },
                "updated_at": {
                    "$ref": "#/definitions/internal.LocalTime"
                },
                "year": {
                    "type": "integer"
                }
            }
        },
...
        "models.EventAlbum": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name_en": {
                    "type": "string"
                },
                "name_no": {
                    "type": "string"
                },
                "time_end": {
                    "$ref": "#/definitions/internal.LocalTime"
                },
                "time_start": {
                    "$ref": "#/definitions/internal.LocalTime"
                }
            }
        },
...
    }
}